requirements.txt
pandas
python-dotenv
groq
gradio
